ui+input: confine SDL to the code that needs it
Two headers pulled SDL in for the sake of a handful of declarations, and
reached far wider than that. The scene module owned the translation of
SDL keycodes into the UI's own key roles, and input/common.h declared
three functions that speak in SDL's own types.

Neither measuring a scene nor reading the input state needs a window
system, so both move next to the code that does. The unit tests have no
SDL headers to find, and now no longer look for them.

The backend header did the same for one function pointer, so it declares
the event type rather than including it, and the backends that read an
event include SDL themselves. What is left is the shell umbrella, which
carries the event and window headers both: the files that reach through
it want one function from it, or nothing at all.
