+++ src/trx/game/replay/test_replay.c
@@ -12,6 +12,7 @@
 #include <trx/game/input/backends/controller.h>
 #include <trx/game/input/backends/keyboard.h>
 #include <trx/game/input/common.h>
+#include <trx/game/input/sdl.h>
 #include <trx/game/lara.h>
 #include <trx/game/lua.h>
 #include <trx/game/random.h>
