+++ src/trx/game/ui/elements/bar.c
@@ -7,7 +7,6 @@
 #include <trx/core/strings.h>
 #include <trx/core/utils.h>
 #include <trx/game/output.h>
-#include <trx/game/shell.h>
 #include <trx/game/ui/draw.h>
 #include <trx/game/ui/helpers.h>
 #include <trx/game/ui/scaler.h>
